+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This changelog covers all three packges, as they are (for now) updated as a whole
 
-## v0.24.1
+## v0.24.2
 
 - Improve resource selector dropdown, show previews, remove dependency #60
 - Add toast notifications #63